Fixed issues in delivery-cost calculation
