+++ src/Sulu/Bundle/Sales/CoreBundle/Resources/public/js/components/item-table/item.form.html
@@ -31,7 +31,6 @@ disabled =  'disabled';
                 data-validation-required="true"
                 value="<%= deliveryCost %>"
                 data-type="decimal"
-                data-validation-min="0.0"
                 data-validation-max="999999999"
         <%= disabled %>/>
     </div>
